added and fix projects

+++ hypertenus.py
@@ -1,6 +1,6 @@
 import math
-A = int(input("what is a: "))
-B = int(input("what is b: "))
+A = float(input("what is a: "))
+B = float(input("what is b: "))
  
 first_half = A*A + B*B
 ans = math.sqrt(first_half)
